refactors API
