+++ server/middlewares/dataValidator.js
@@ -1,7 +1,64 @@
 import 'babel-polyfill';
 import db from '../db/connection';
+import validationHelper from '../helpers/validationHelpers';
 
-class UpdateValidator {
+class ImproperValues {
+  static improperUserData(req) {
+    const {
+      fullname, email, phoneNo, password,
+    } = req.body;
+
+    const fnameTest = /^[a-zA-Z]+? [a-zA-Z]+?( [a-zA-Z]+?)?$/.test(fullname);
+    const emailTest = /[-.\w]+@([\w-]+\.)+[\w-]{2,20}/.test(email);
+    const phNoTest = /^\d{10,20}$/.test(phoneNo);
+    const passwordTest = /.{7,}/.test(password);
+
+    const improperValues = [];
+
+    if (!fnameTest) improperValues.push('Improper name pattern. There should be a space between first and last name. E.g: \'John Smith\'');
+    if (!passwordTest) improperValues.push('Password too short. Should be at least 7 characters');
+    if (!phNoTest) improperValues.push('Invalid phone No');
+    if (!emailTest) improperValues.push('Invalid email');
+
+    return improperValues;
+  }
+
+  static improperParcelData(req) {
+    const {
+      pickupAddress, destination, pickupTime, parcelDescription, parcelWeight,
+    } = req.body;
+
+    const timeTest = /^\d{4}-((1[0-2])|(0[1-9]))-((3[0-1])|([0-2][0-9]))T(([0-1][0-9])|(2[0-3])):[0-5][0-9]$/.test(pickupTime);
+    const pAdTest = /.{15,}/.test(pickupAddress);
+    const dAdTest = /.{15,}/.test(destination);
+    const pDeTest = /^.{3,40}$/.test(parcelDescription);
+    const pWeightTest = /(kg)/.test(parcelWeight);
+
+    const improperValues = [];
+
+    if (!timeTest) improperValues.push('Improper date-time format or invalid date. Pattern should follow: YYYY-MM-DDThh:mm and date/time must not be behind present');
+    if (!pAdTest) improperValues.push('Pickup Address not detailed enough');
+    if (!dAdTest) improperValues.push('destination not detailed enough');
+    if (!pDeTest) improperValues.push('parcel description not detailed enough or too long. Min. Length:3, Max. Length:40');
+    if (!pWeightTest) improperValues.push('parcel weight unit must be in kg');
+
+    return improperValues;
+  }
+}
+
+class DataCreationValidator {
+  static userDataValidator(req, res, next) {
+    const dataKeys = ['fullname', 'email', 'phoneNo', 'password'];
+    validationHelper(req, res, dataKeys, ImproperValues.improperUserData, next);
+  }
+
+  static parcelDataValidator(req, res, next) {
+    const dataKeys = ['pickupAddress', 'destination', 'pickupTime', 'parcelDescription', 'parcelWeight'];
+    validationHelper(req, res, dataKeys, ImproperValues.improperParcelData, next);
+  }
+}
+
+class DataUpdateValidator {
   static async cancel(req, res, next) {
     const { parcelId } = req.params;
     try {
@@ -9,14 +66,12 @@ class UpdateValidator {
       if (!result.rows[0]) return res.status(404).json({ message: 'Order not found' });
       if (result.rows[0].status === 'delivered') return res.status(400).json({ message: 'You cannot cancel an already delivered order' });
       next();
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
     }
-
   }
 
-  static async changeStatus(req, res, next) {
+  static async status(req, res, next) {
     const validValues = ['recorded', 'dispatched', 'in transit', 'delivered'];
     const { status } = req.body;
     const { parcelId } = req.params;
@@ -27,8 +82,7 @@ class UpdateValidator {
       if (!status) return res.status(400).json({ message: 'invalid request, new status not provided' });
       if (!validValues.includes(status)) return res.status(400).json({ message: "Invalid status value. Value must be either 'recorded', 'in transit' or 'delivered'" });
       next();
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
     }
   }
@@ -41,13 +95,12 @@ class UpdateValidator {
     try {
       const result = await db.query('UPDATE parcelOrders SET status= $1, receivedBy= $2, receivedAt= $3 WHERE parcelId= $4 RETURNING *', [status, receivedBy, receivedAt, parcelId]);
       res.status(200).json(result.rows[0]);
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
     }
   }
 
-  static async changeDestination(req, res, next) {
+  static async destination(req, res, next) {
     const { destination } = req.body;
     const { parcelId } = req.params;
     try {
@@ -57,8 +110,7 @@ class UpdateValidator {
       if (!destination) return res.status(400).json({ message: 'invalid request, new destination not provided' });
       if (!/.{15,}/.test(destination)) return res.status(400).json({ message: 'destination not detailed enough' });
       next();
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
     }
   }
@@ -74,14 +126,13 @@ class UpdateValidator {
     try {
       const result = await db.query('UPDATE parcelOrders SET status = $1, presentLocation = $2 WHERE parcelId= $3 RETURNING *', [status, presentLocation, parcelId]);
       res.status(200).json(result.rows[0]);
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
       next();
     }
   }
 
-  static async changePresentLocation(req, res, next) {
+  static async presentLocation(req, res, next) {
     const { presentLocation } = req.body;
     const { parcelId } = req.params;
     try {
@@ -91,11 +142,10 @@ class UpdateValidator {
       if (!presentLocation) return res.status(400).json({ message: 'Invalid request, present location not provided' });
       if (!/[\dA-Za-z]{1,20}/.test(presentLocation)) return res.status(400).json({ message: 'invalid location or location length too long' });
       next();
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
     }
   }
 }
 
-export default UpdateValidator;
+export { DataCreationValidator, DataUpdateValidator };
