--- server/model/parcelData.js
@@ -1,12 +1,0 @@
-export default [
-  // sample data
-  {
-    userId: 'ayzay', parcelId: 1, pickupAddress: '8. Babudu close, Isashi', destination: '8. Muibi st. Ogba', pickupTime: '2018-11-06T22:07', parcelDescription: 'Bag of Garri, carton of Indomie', status: 'in transit',
-  },
-  {
-    userId: 'nduka', parcelId: 2, pickupAddress: '8. Babudu close, Isashi', destination: '10. Straight rd, Albany', pickupTime: '2018-11-09T10:17', parcelDescription: 'Samsung Television', status: 'delivered',
-  },
-  {
-    userId: 'ayzay', parcelId: 3, pickupAddress: '14. wetheral rd, Owerri', destination: '29. Akanihu close, Ndianiche', pickupTime: '2018-12-06T19:07', parcelDescription: 'Jewelries', status: 'received',
-  },
-];
