--- server/model/userData.js
@@ -1,9 +1,0 @@
-export default [
-  {
-    fullname: 'Ogbonna Ezenwa', userId: '0874yhshjjdhhjjwrjhegr', email: 'ezenwaogbonna1@gmail.com', phoneNo: '08140064376', password: 'ozone4real', isAdmin: true,
-  },
-
-  {
-    fullname: 'Ogbonna Nduka', userId: 'ugdyeiowpwytwwi', email: 'ndukadenniece@gmail.com', phoneNo: '08084900474', password: 'denniece456', isAdmin: false,
-  },
-];
