+++ server/routes/parcels.js
@@ -1,26 +1,35 @@
 import { Router, json, urlencoded } from 'express';
-import validateOrder from '../middlewares/parcelOrderValidator';
-import UpdateValidator from '../middlewares/parcelUpdateValidator';
+import { DataCreationValidator, DataUpdateValidator } from '../middlewares/dataValidator';
 import ParcelController from '../controllers/parcelController';
 import Auth from '../middlewares/auth';
+
+const {
+  getAllOrders, getOneOrder, createOrder, cancelOrder,
+  changeDestination, changePresentLocation, changeStatus
+} = ParcelController;
+const { userAuth, adminAuth } = Auth;
+const { parcelDataValidator } = DataCreationValidator;
+const {
+  status, destination, presentLocation, delivered, cancel, inTransit
+} = DataUpdateValidator;
 
 const router = Router();
 
 router.use(json());
 router.use(urlencoded({ extended: false }));
 
-router.post('/', [Auth.genAuth, validateOrder], ParcelController.createOrder);
+router.post('/', [userAuth, parcelDataValidator], createOrder);
 
-router.get('/', [Auth.genAuth, Auth.adminAuth], ParcelController.getAllOrders);
+router.get('/', [userAuth, adminAuth], getAllOrders);
 
-router.get('/:parcelId', Auth.genAuth, ParcelController.getOneOrder);
+router.get('/:parcelId', userAuth, getOneOrder);
 
-router.put('/:parcelId/cancel/', [Auth.genAuth, UpdateValidator.cancel], ParcelController.cancelOrder);
+router.put('/:parcelId/cancel/', [userAuth, cancel], cancelOrder);
 
-router.put('/:parcelId/status', [Auth.genAuth, Auth.adminAuth, UpdateValidator.changeStatus, UpdateValidator.inTransit, UpdateValidator.delivered], ParcelController.changeStatus);
+router.put('/:parcelId/status', [userAuth, adminAuth, status, inTransit, delivered], changeStatus);
 
-router.put('/:parcelId/destination', [Auth.genAuth, UpdateValidator.changeDestination], ParcelController.changeDestination);
+router.put('/:parcelId/destination', [userAuth, destination], changeDestination);
 
-router.put('/:parcelId/presentLocation', [Auth.genAuth, Auth.adminAuth, UpdateValidator.changePresentLocation], ParcelController.changePresentLocation);
+router.put('/:parcelId/presentLocation', [userAuth, adminAuth, presentLocation], changePresentLocation);
 
 export default router;
