+++ server/routes/users.js
@@ -2,11 +2,13 @@ import { Router, json, urlencoded } from 'express';
 import UserController from '../controllers/userController';
 import Auth from '../middlewares/auth';
 
+const { userAuth } = Auth;
+const { getAllUserOrders } = UserController; 
 const router = Router();
 
 router.use(json());
 router.use(urlencoded({ extended: false }));
 
-router.get('/:userId/parcels', Auth.genAuth, UserController.getAllUserOrders);
+router.get('/:userId/parcels', userAuth, getAllUserOrders);
 
 export default router;
